Distributing java v1.6.0_23
Based on info from the group revision 22 was provided key crypto
handshaking that was necessary for a more secure federation.

------------------------------------------------
email from Stephen Pascoe:
------------------------------------------------
When doing any software upgrade to ESG Federation servers I would
encourage everyone to upgrade their JVMs to 1.6 update 22.  This
update reintroduces TLS renegotiation using RFC-5746.  Although we
will not be able to use renegotiation until all clients and servers
support RFC-5746 it is worth putting the upgrades in place as soon as
possible.  See this link for details:

  http://download.oracle.com/javase/6/docs/technotes/guides/security/jsse/JSSERefGuide.html#phased

Openssl 0.9.8m+ supports RFC-5746 so we should also upgrade that --
unfortunately not always easy, for instance SuSE Enterprise 11 doesn't
have an RPM for this.  NOTE: we should *not* upgrade to openssl-1.0.
Phil believes there is a significant change in truststore format that
would cause problems.
------------------------------------------------