Shoring up esg node manager upgrade procedure...
We want to overlay the installation upgrade over the current one in
the usual case.  Unless we --force an upgrade.  We should rely on the
initial version checking to be our gatekeeper and not have to do it
any further in the code (this was an old pre version check vestage:
$no_new_update).
And a quick echo to let me know how the propery file is being
manipulated (edited or copied back into place).