Won't ask when removing tomcat examples directory unless using --verbose
I am not too sure about this.  I want to be cautious so that anytome
"rm" is being called the user has the opportunity to inspect the
directory just for safety.  Instead I will display the absolute path
of the directory to be removed so at least there is a visual record of
it.  If you use the --verbose flag it will ask you directly.