modified return value of self_verify in the context  of --check flag.
Also, some verbiage change in help for the --check flag
The purpose of this is so that when you do

%> esg-node --check c

if the script is not up to date it will return a non-zero value
to further indicate that the current script is out of date.  This is
needed because it provides information to the caller regarding how
up-to-date the resident script is.  This makes it useful when put in a
crontab scenario where the return value of the call can be used to
trigger email notification of cronjob task etc.

If you do

%> esg-node --check u

The code will always update to whatever is on the distribution server.
This keeps the resident script in sync with the distibution server's.