Bug #20 - [RESOLVED] Increase PermGen for Tomcat application
http://esgf.org/bugzilla/show_bug.cgi?id=20

Bug statement:
After deploying the esgf-security war file, the Tomcat application
crashed
because of insufficient PermGen. It is suggested that the datanode
installation
start allocating more PermGen to Tomcat as such:

java_opts=${JAVA_OPTS:-"-Xmx2048m -Xms1024m -XX:MaxPermSize=128m"}

The script would already set this value, but only for the "IDP" node
type.  To address this bug I made it so that this value is set for all
node types.

I suspect it would be prudent to pair down this to only allocate this
memory for the specific types this is needed for.