fixed problem with --gridftp-config flag args
The problem was that init was being called too often and resetting
things. Also, the call to source the environment /etc/esg.env should
not also be so micro managed.  It should take place once.  I put the
call to init at the TOP of the main().  It seemed to properly fix the
--gridftp-config issue - of the value not being sent and pushed
forward into the globus startup sequence.

-gavin