Allow for different keystore password - added arg flags at command line
Users can set values that will override the uber password w.r.t. keystores

--dname
--keystore-password
--keystore-alias
--keystore-file

Have utility method to transform from the 'standard' DN format to the
'java-style' DN format - as_javastyle_dn().  This is pretty much the
reverse functionality of the code within extract_keystore_dn().  The
latter pulls the dn directly from a given keystore prior to performing
the dn transformation from 'java-style' to 'standard'.

The user is given a prompt for DN value with a default and ability to
accept or reject it and go into the keytool invocation that will ask
for the DN information.  The keytool uses this dn to generate the slef
signed cert keypair that is embedded in the keystore.