The last portion of the initialization split off so that it is not root dependent.
There are some tasks at the end of initialization that make it so that
you need to be root.  i.e. making directories in certain places etc.
The modification here is to move those actions to just prior to
installation.  Also the variables used are defined in the init, which
gets called at the top of the main.  As such it enforces the general
variable setting workflow as

1) Read from environment
2) Get set from command flags
3) Get read in from configuration property file

Many environment variables get saved out ultimately as properties.
So that subsequent runs will have those values persisted and used
unless usurped by the environment or command flags.