Lots of Certificate gymnastics
Added function to extract DN from openssl x509 output and turn them
into out "standard" format.

At the end of a successful key-pair install setting the
esgf.properties node.dn value to the new DN

Putting a conditional around cert chain verification to make sure
there is a chain in the first place.  This should always be the case
and most certainly is the case when it comes to signing anything
real.  Because anything real depends on a CA signing which means you
have at least one cert in the chain, the signing CA's.  But in the
self signing situation i.e. there is no chain, I want to make sure I
don't panic anyone with an irrelevant error message.