Fix always-failing postgres user check
The check wasn't being made with user "postgres", and no permissions
have been set up for user "root", so this always fails, and the
resulting psql error message interferes with the password gathering
stage, making it much more likely that there will be a password
mismatch in the following step.  This change forces -U postgres on the
check, which appears to resolve the issue.

Fixes: #72