Can you update the references to this file on rainbow - authz_esgsaml_callout (09-02-2010 - 07-18-2011)
Hello Gavin,

Can you update the references to this file on rainbow and in the
install scripts?  It's a one line change that disables calling the
authorization server on writes.  Obviously this doesn't make sense
because the authorization service will never have mappings for
destination files that it doesn't know about.

http://www.mcs.anl.gov/~neillm/esg/authz_esgsaml_callout-07-18-2011.tar.gz

I found this bug in testing with NCAR, but it affects all GridFTP
servers out there (which leads me to believe that means there are
either none, or are running an even older version of the module than
the last).  The call to the auth service on writes was added in the
last version so NCAR could test something, but there was no follow-up
on it, and it shouldn't have been permanent.

I've pre-emptively updated the docs from:

http://rainbow.llnl.gov/dist/globus/gridftp/authz_esgsaml_callout-03-09-2010-2.tar.gz

to:

http://rainbow.llnl.gov/dist/globus/gridftp/authz_esgsaml_callout-07-18-2011.tar.gz

thanks,
-Neill.