Aware of how to check for process only bound to particular IP
Prior to this commit, the test for tomcat would check for anything
bound to the configured ports.  This did not discriminate by IP
address.  Now we will not only check for ports being bound but also
what IP address they are being bound to.

This was an outstanding issue and further underscored by Luca...

----------------
Hi Gavin,

        I wonder if you could make a modification to the esg-node
startup script. Right now, before starting, the script checks wether
Tomcat is already running on port 80, and if it finds a process on
port 80, it aborts. This is all good, but the testing command should
filter for processes that run on the specified IP address for that
node, for example:

local procs="$(lsof -Pni TCP:$ports | tail -n +2 | grep LISTEN | grep 172.24.254.126 | awk '{print $1}' | sort -u | xargs)"

      The reason is that on my NOAA server I already have another
process running on another IP address on port 80, and right now it
interferes with starting of the ESGF node. Could the script be changed
to be more specific ? Pretty please ?

thanks, Luca
----------------

No worries Luca, this should do it.
-gavin