LAS to version 7.3.2
As per Roland's email...

------------------------------
Hi,

I have packaged and tested a new release of LAS for use with ESG.

Please update the install scripts to pull this tar file:

ftp://ftp.pmel.noaa.gov/pub/las/las-esg-v7.3.2.tar.gz

for LAS installation.

This release fixes a couple of usability issues specific to ESG.  Data
plots now automatically load when you land on the LAS page after linking
from the web-fe search (instead of waiting for you to push the "Update
plot" button) and all the variables in a data set are collected into a
single LAS category which makes them accessible from the drop down
selector and allows you to do property-property plots of those
variables.  Plus the release has all of the LAS bug fixes that have
accumulated since the previous release.

Roland
------------------------------

AND...

After the new LAS is installed you must either restart your Tomcat
container or hit the reinit URL of your LAS to force it to re-read its
configuration.

The details of the admin and reinit interface are here:
http://ferret.pmel.noaa.gov/LAS/documentation/installer-documentation/las-administration

but just restarting the tomcat will be easier for testing.  We don't
actually recommend upgrading and LAS while tomcat is running.
Roland
------------------------------