check_version_helper() now uses check_version_atleast() and check_version_between()
This arrangement reverts the previous commit changing the order of
arguments in check_version_atleast() and check_version_between() so
that they will be consistent with the autodetectable format used in
check_version_helper(), despite the fact that the syntax may be
independently slightly less intuitive.

In all three functions, the version to be tested is now the first
argument.

The following functions were modified to make use of the new
check_version_helper() syntax:

  check_version()
  check_module_version()
  check_webapp_version()
  check_app_version()
  check_version_with()