Including release value in script version output
Now will grab the version value from the git tag and include it in the
--version output.  Also taking a feather from Zed's cap I am creating
"random" temp files as I go through the sed editing machinations.
(this is a bit ugly but necssary to make it work on both mac and unix)
close your eyes and hit enter! :-) It works!! :-)