Using the --no-globus option to turn off writing of /etc/grid-security/host{cert,key}.pem files when installing a new keypair
As it stood, everytime you installed a new keypair into tomcat's
keystore (and subsequently including it in the truststore) you would
use that same pair as the keypair for turning on globus.  Now you may
leave your keypair for turning on globus (files
/etc/grid-security/host{cert,key}.pem) in tact while installing only
into tomcat's keystore and truststore.  Indeed it is usually the case
where the keypair in question would be used for both.  Especially if
the keypair is signed by a CA in the browser cartel CA group.

Additonally I have commented out a call for setting the node's dn.
Right now the Node's DN is the DN of the keypair (as disscussed
above).  It seems to me that the DN of the simple CA should possibly
be reflected since it is the cert needed when logging into the system
via myproxy as well as used by gridftp for file transfers. The simple
CA's cert seems to be most appropriate for the title Node DN.

Or perhaps I should create another entry and have both DNs present?

This is a managerie of certificates that is really difficult to keep
up with.  The way we are using myproxy and the simpleCA is unwieldy
and I fear will not scale!  The good thing is that you only need a
simpleCA for every IDP in the federation, which should be far fewer
in number than any other node type. (whew)