Ability to not download files from dist server in order to test local mods
This is a feature that I should have added a while back.  It allows
developers to make changes to local install and configuration files
and keep them in tact while they continue to install and exercise it,
without worry of it being overwritten by the distribution server's
version.

I will not advertise this option as we want to be able to assert
installations' use of pristine distribution server versions of files.

This only works for files fetched via the checked_get function.
Files gotten directly via wget or curl are not affected.

-gavin