calling function from the node-manager install script - makes sure whitelist files are appropriately commissioned
For some cases it has been seen that the derivative whitelist files
are being owned by root and thus when the tomcat owned process
attempts to write to these files there is an exception due to
permission denied.  So here, as root, we touch the files and then
chown and chmod them appropriately.

Cross ref this commit with the node manager project's commit
containing this function definition.

http://esgf.org/esgf-node-manager.git:
commit c75c587ae3039c97bad35441705bd210ee833c86
Author: Gavin M. Bell <gavin.max.bell@sixthcolumn.org>
Date:   Sun Nov 13 23:29:46 2011 -0800