Fixed up _add_to_path function...
The function generates a string that is the instruction to generates
the export variable statement.  The key feature of this function is
that it dedups anything you add on to the base path.

use it like this:

alpha=beta

add_to_path FOO three four five six seven eight three ${FOO} ${alpha} one

Gives as output the string
export FOO=three:four:five:six:seven:eight:beta

Notice it did not repeat "three" in the path it constructed
Calling it as a process substitution directly exports the var into the
environment.