Now have each install function write apporpriately to PATH...
Without having to worry about making the path too long, etc.
This function works well for deduping path additions.
(better than pathmunge)

Still keeping the a-priori setting of the PATH and LD_LIBRARY_PATH
because it crates a source-able env script that does not contain the
user's specific path but evaluates $PATH and $LD_LIBRARY_PATH when
this script is sourced.  I went back and forth on this and in the end
I decided to keep what was there, and only augment the setting of the
path vars so they may be done in the installing function.

All praises to _add_to_path! :-)