Do not replace host keypair when installing new SSL keypair
The lesson learned is that host keys and ssl keys are not the same
beast. The installer is currently setup to properly manipulate and
install SSL certificates.  These are certificates that live in the
tomcat keystore (and truststore).  This task is done.  Host certs are
of a different ilk and are used by Globus to secure services on a
*host*.  These are not the same as SSL certs and thus I cannot copy
them into /etc/grid-security/host{cert,key}.pem and have Globus
services work.

Thus I am commenting out the last portion of the SSL certificate
installation gymnastics, which copied the files into the globus
location.

TODO:
Generate CSR using globus tool that is of the HOST type.
(suitable for use for Globus services)