Oops, fixed bug with sed call to look only for processes that are open or bound to my designated port
Indeed, the goal is to do an lsof command but pared down to only those
socket processes that are bound to all interfaces or only to the IP
specified for this node.  I forgot the "-n" on the sed command that
does this pruning.  Alternatively I could not use the /p.

This should do it.

Thanks Luca for bringing this to my attention:

> Hi Gavin,
> could you insert the additional "grep" instruction in the esg-node script:
>
> local procs="$(lsof -Pni TCP:$ports | tail -n +2 | grep LISTEN | grep ${esgf_host_ip} | sed 's/\(\*\|'${esgf_host_ip}'\)/\0/p'  | awk '{print $1}' | sort -u | xargs)"
>
> This would take care of checking the Tomcat ports when there are multiple interfaces on the same machine.
>
> thanks, Luca

We don't need an additional grep, just to get the sed doing what it
was supposed to do :-)

-gavin