Custom patched gsoap and forcing a globus gtp-build of the gss_assist_patch... (see diff for rationale)
Neill Miller made a good patch to gsoap, as newer gcc compilers are
more strict and will not let shenanigans pass :-).

Also, there seemed to be a strange bit of behavior such that the
gss_assis patch somehow gets interfered with, we speculate it may be
the subsequent BDM configuration process.  Suffice it to say, it works
with after the fact when the same gtp-build installation of the patch
is done, all works fine.  So making the call again (it is really
quick) after all the gridftp related code has been installed.
Hopefully this will address the issue.