Edit all web.xml files to reflect new authorizationService endpoint
When a user changes idp peers, they usually are also changing the
authorization service endpoint as well.  This value is specified in
the web.xml of all web applications that refer to it.  This change
makes it so that as the index peer is repointed the web.xml files are
also re-pointed to that node/host as well.

Note:
The variable is authorization_service_root specified in the
esgf.properties file.  If that value is set this is the value that is
used when editing the web.xml file.  If this value is not set the
default value is the esgf_idp_peer value.

see:
     esgf-installer/esg-node
     esgf-installer/filters/esg-security-tokenless-filters
     esg-dist/bootstrap/filters/esg-security-tokenless-generic-filters.xml