Making sure publisher is set to P2P index peer
The template that is generated for the publisher has in it the URL for
the older gateway publishing service.  This is not suitable for the
P2P system.  So we are making sure that upon installation of the P2P
node that the publisher is pointing to the index peer node we
indicated from when the questionnaire was answered in the beginning.