setup_postgress() explicitly checks for the version of /usr/local/pgsql/bin/psql
Since this function only runs when the database is NOT externally
managed, checking for the version of whatever psql is in the path is
not appropriate, and causes the database install to fail if there is a
system version of the postgresql client available.

This commit also adds a debugging message to the top of the script,
and slightly changes the 'return if successful' logic to prepare for
the addition of extra sanity checks later.