Made change requested by Roland...Removing a property pair in  config file...
His email: on Feb 27 @ 2:13pm
-----------------------------------
Gavin,

There is a file (esgcet_las.xml)  that gets created as part of the
node install of LAS that needs to be changed.  As far as I can tell,
that file is not a part of my distribution tar bundle, but is created
by the node script itself.

The contents of the file are something like this:

<datasets>
    <dataset name="ESGF Data" src_type="THREDDS"
    src="http://YOURTHREDDSSEREVER/thredds/esgcet/catalog.xml">
        <properties>
            <property_group type="addXML">
              <property>
                  <name>esg</name>
                  <value>true</value>
              </property>
              <property>
                  <name>category</name>
                  <value>true</value>
              </property>
           </property_group>
       </properties>
   </dataset>
</datasets>

The part in bold needs to be removed so it looks something like this:

<datasets>
    <dataset name="ESGF Data" src_type="THREDDS"
    src="http://YOURTHREDDSSEREVER/thredds/esgcet/catalog.xml">
        <properties>
            <property_group type="addXML">
              <property>
                  <name>esg</name>
                  <value>true</value>
              </property>
           </property_group>
       </properties>
   </dataset>
</datasets>

How can I make this happen and how can we work it so this file gets
fixed on existing installations?

Thanks,
Roland
-----------------------------------
Consider it done Roland :-)