Sanity check on certificate result to be sure has suitable Subject (DN) specified
There seems to be a situation such that an installation takes place
and a CSR is generated without a Subject: / DN line.  This is thus
invalid for our system to use to generate a certificate.  So this is a
small sanity check for that subject, along with some instructions on
what to do to surmount this problem.