Enforce schema nameing convention with "esgf_" prefix
In order to adhere to esgf schema naming conventions I have set it up
so that esgf schemas may be indicated with or without the "esgf_"
prefix.

Run this to see what I mean:

f=esgf_node_manager
[ "esgf" = "${f%%_*}" ] && echo "name = ${f#*_}" || echo "name = $f (clean)"

f=node_manager
[ "esgf" = "${f%%_*}" ] && echo "name = ${f#*_}" || echo "name = $f (clean)"

Both are accepted now.  The eases the burden on the caller, now no
longer having to remember if the schema argument took the prefix or not.