Hard setting of HOME to /root to satisfy new globus installation assertion
It has been asserted that the home for the simpleCA installation must
be in the root user's home directory specifically /root/.globus.  The
globus programs use $HOME to determine where to perform installations.

To accommodate this we are setting the HOME var explicitly to /root.
To aide in the migration of older installs that used $HOME in the same
way, but whose value was set to the home directory of the created
installation user - there is the _mv_simpleca_to_root() function.
This function effectively moves the simpleCA installation from the original
location (if found) to the new location and archives the original.

NOTE this entire script is run from esg-node in a sub-shell.
Because of the screwing around with the $HOME var (unfortunate) we
want to isolate this script from poluting the environment of the
calling environment.  ALWAYS CALL THIS PROGRAM IN A SUBSHELL!!!

See commit: af677c14587e33fb8952b4baa023cb94fcac0e54 -- esg-node

-gavin