Along with simpleCA migration be sure to reconfigure mapapp...
When a simpleCA migration is performed it also requires a modification
in the myproxy-certificate-mapapp script used in the generation of
ephemeral myproxy certificates.  The modification points the mapapp
file to the now migrated (moved) simpleCA installation.

A migration is noted by a "boolean" value of 1 (true) given to
"private" variable: _performed_simpleCA_migration.

When this, _performed_simpleCA_migration value, is passed to the
fetch_myproxy_certificate_mapapp function - as true - it forces a
re-fetch of the template and thus a re-edit of the template file to
have the new path to the migrated (moved) simpleCA, now in
/root/.globus

While I was at it I made all the edit and fetch functions take a
"boolean" argument indicated if it should be "forced" or not.

-gavin