Bug fix: version checking helper function - not comparing as it should
when the vesion comparison function _versoin_segment_cmp() was faced
with comparing

[[ 9 > 10 ]]

it would result in a TRUE evaluation.  This is in correct.
So now checking to see if alpha or numeric values are being
compared. If it is a purely numeric comparison then using the (( ))
bash syntax for comparing numbers - in which case

(( 9 > 10 ))

Is properly evaluated as false.
Also, if there is no value at that segment the default is set to 0
(zero).

If we detect that it is an alpha comparison we do what was done before
using [[ ]].

This came up as it was checking for the new esgcet version which
compares

2.9.1 with 2.10.1