NO longer disabling end-user commands.... relying on callout mechanism
Neill,

I made the modifications... however...
What are the security implications for this?

As it was the following commands were disabled...

-disable-command-list APPE,DELE,ESTO,MKD,RMD,RNFR,RNTO,RDEL,STOR,STOU,XMKD,XRMD,CHMOD"

Now, what is the default behavior now?
What commands that the user can do?
Should we still disable things like CHOMD and some others?

On 6/6/12 2:52 PM, Neill Miller wrote:
> New Authz Saml Module ...
>
> http://www.mcs.anl.gov/~neillm/esg/authz_esgsaml_callout-06-06-2012.tar.gz
>
> If you could add that to rainbow and post the script changes we discussed (just this file reference, and commenting out the disable_command_list line and setting it instead to be disable_command_list=""), I can test the globus --devel installer tomorrow.
>
> thanks,
> -Neill.
>