Bug fix-ish: Make sure that the node responds to "hostname --fqdn" command
Ran into weird issue that if the node does not have a fully qualified
domain name that "hostname --fqdn" returns a "host unknown" response.
So we have to make sure that the host is given a fully qualified
domain name value in (on CentOS) /etc/sysconfig/network for the
hostname key.  It is a strange corner case that I found.

It is strange because straight "hostname" works (and gives the single
name value of the hostname).

By the way to make this take effect after changing the "network"
file... you have to *reboot* the machine.

So this check will do an up front check before things start to cascade
out of control as various parts of the code looks to "hostname --fqdn"
for a value... (sort of)