Defensive installing for users that should install COMPUTE
There was a small issue with folks that perform installations without
turning on the compute bit, yet are publishing records that have been
configured to advertise that they have 'computable' information.

If the esg.ini sets "las_configure = true" on any project then the
system will alert them that they should probably install the compute
type.  They are prompted and if they choose to they will be able to
install the compute type to support what they advertise that the user
can do with that data.

This came up during the ESGF F2F when fielding an issue with Roland.
:-)