Wrap use of md5sum such that it works for Linux AND Darwin machines
On macs the defacto file is "md5" on Linux it is "md5sum".  So wrapped
the call to those programs in a function that, based on what os you
are on and will select the appropriate program.  Also the output of
the program md5 (on the mac) is re-written to look the same as the
output of md5sum found on linux machines.

Additionally the output of the delegated md5 call is trimmed and
written to the *.md5 file (not the actual temp file that was
checksummed)