Have CMake build zlib based on patch value of currently installe zlib...
There is an issue where uv-cdat needs to build it's own zlib so it is
clear on what it is using.  If there is zlib already on the machine
uv-cdat will build its own ostensibly equivalent version of zlib that
it can use independent of what is on the system.  If there is no
suitable libz patch version of 1.2.x where x = 3,5,7 then the default
is to build libz at patch version 3 (i.e. 1.2.3).

Libz is needed for lib xml which is used most notably by the ESGF
Publisher.

-gavin