Put /usr/local/ father down PATH
The issue is that binaries being looked for are being picked up in
places that THIS installer may not have put them.  The first thing is
to look for files that THIS installer may have created before looking
for files that *another* agent may be put in place.

Ex: curl... /usr/local/curl is found before /usr/local/curl/bin/curl

The latter is the one ESGF installs the former can be found on the
system.  This will make version checking for curl always be off
w.r.t. what the ESGF installer has installed.