Deeper checking for "ROOT" webapp...
It is not good enough that there is a ROOT directory in webapps.
Because when installing a new tomcat there is indeed a ROOT app that
is there (esp tomcat v7).  So we want to do a little deeper probing at
the file to see if we see the word REFRESH.  Yes, yes... I know  it is
a weak test.  I perhaps could have come up with something more
definitive... maybe a sexy regex... I'll put it on my list of TODOs
;-).

Suffice it to say that if we check and see the ROOT webapp and further
see that it does not have "REFRESH" - we back up the ROOT webapp and
then proceed with installing our own ESGF "ROOT" webapp. :-)