Zlib shennanigans...
So before we tried to detect the system zlib and build the SAME
version.  It turns out that when we did that there was not "version
information" evident in the uv-cdat built libary.  So now if we don't
build it at all and use the system one, then other programs that use
zlib won't run into it from tracing the LD_LIBRARY_PATH and finding
the uv-cdat built one (w/o version number) and instead will find the
system zlib under /usr/lib${word_size} :-)