(globus) use different post -subject flag... s/-force/-days/
After the -subject flag, if there isn't another flag the script
setup-simple-ca hangs.  Eric B. advises not to use -force flag as it is a
bit heavy handed so will use the more benign -days flag setting it to
the default 5 year value. Thanks again Eric B. for the input.

email:

Cool, looks good, with the possible caveat that we may not want to be
using -force (in case someone invokes this code in error, we don't
really want them to accidentally overwrite their CA).

So maybe we want "-days 1825" (which is the default anyway) or something instead.

Eric