Should address the "cdat hanging" issue...
In the context of ESGF UV-CDAT will never log your usage information.
Upon installation and the initial call to load the cdms library the
user is prompted if they want to do ANONYMOUS logging or not.  It is
this question that the user is being prompted for that is not being
echo'ed to the terminal and thus looks like or ostensibly equivalent
to "hanging".

By setting the environment var UVCDAT_ANONYMOUS_LOG=OFF you are not
prompted thus won't get "hung"