Qualifying that the IP provided should be more specifically a routable IP (implied by "public")
It is not so much that it needs a "public" IP address more like a
routable IP address.  The node manager and the dashboard/desktop only
deal with routable IP addresses as they are more concerned with public
node interactions. This is built that way so that we give you info on
the world at large that can be "seen" instead of an intranet behind a
faux IP scheme.

There are three address ranges designated as private as per RFC 1918:
10.0.0.0 … 10.255.255.255 (10/8) - (Class A network that start with 10.)
172.16.0.0 … 172.31.255.255 (172.16/12) - (Class B networks that start with 172.16.)
192.168.0.0 … 192.168.255.255 (192.168/16) - (Class C networks that start with 192.168.)