Get the esg-functions file if it is older than the esg-node script
We want to make sure that the functions file being used is as
up-to-date as possible, thus it must be at least as old as the install
script itself.

So now... if the esg-node file is newer than the functions file... the
functions file gets freshly fetched. (otherwise just use what is
there)