No longer using publisher egg - installing w straight up setup.py on source
Dispite Bob's best intentions, without his care and feeding of
building eggs - we could not justify continuing to use them as that
mechanism is python culturally out of favor and adds issues when
installing that we could not easily surmount another way.

The result is the same... And now we can directly install from the
source in git.

P.S.
And Charles can't stand eggs! :-)

-gavin