do not update THREDDS web.xml if not a data node
Fix a line that is breaking "--update" with an index-only node. Even if for the call to get_webxml_file there is no condition on $((sel & INSTALL_BIT)), it should still be conditional on the data component having been selected. If not a data node, it can do nothing useful, and in fact breaks the script because it tries to download the web.xml file to a non-existent local directory.