Force Globus installation to Y at all times
This gets run twice during a fresh install for --type all (once for
gateway, once for datanode), but the second time the prompt will
default to N because it detects an existing Globus installation.  This
has to be forced Y, or a number of packages will not get installed and
the Globus user will not get created.