Move check into set function and check after reading node type
The node type always is read from the node type file, even when it
is specified at the command line via --type. For this reason it is
checked before being written and after been read. This assures that
it was not edited in between and that the write was correct.