Make reading from a INI config file and getting a property disctint
A property is a value that is relevant to the installer and the
eventual configuration of the node. They fall into many different
categories. Some have a finite set of choices, some may be None,
some may need to match a regex, some require prompting the user,
others may not, etc. This allows for the ability to obtain values
while maintaining restrictions on that value as well as where to
get the value from.