Add log4j.jar installation
This system for installing these jar files seems highly unstable.
Should it not be as simple as pull all the needed .jar files from
the distribution mirro and copy them into their respective
locations? Why are they going into orp and then later being copied
to thredds? Also, the names of the files are hardcoded? What if the
version in the mirror changes or a jar is no longer needed? Unless
I am missing something this could use a significant overhaul.