Use static pg_hba.conf file
The pg_hba.conf did not contain entries for IPv4 connections. Also,
it is a static file and should not be rewritten, buy rather copied
into place. The new pg_hba.conf file is the one being used by the
2.x installer and is also the default generated by postgres. It
isn't really needed for us to do this step as postgres will put
this exact file in place if a pg_hba.conf is not already there.