Fixed incomplete database setup
The "postgres_create_db()" function incorrectly returns a non-zero
value when the database is created successfully.  This in turn
causes the security and node-manager scripts to terminate the rest
of their database setup.  On the IDP node, this was mitigated by the
fact that the node-manager is setup (and then later removed) first
where the database is initially created and then the security
setup script was able to run to completion.