Further implement and test the download and extract method
It appears the desired affect of extracting tarfiles is to move the
root tar directory to the specified location. This had been done
via symlinks in the past. This approach simply moves the root tar
directory to the specified location as this is not a typical use
case for symbolic links.