Restructure component configuration and specification
The components, their attributes, what method installs them, and
what node type they belong to are all specified in a single location.
To make it easy to assign components to method types and component
types as well, the central file is a .py file that imports the class
types of methods and components. Lightweight components, that is
components without a need for anything beyond what the method does,
can be assigned a generic component type. Heavyweight components,
those with pre- and post-install steps, can be assigned their
specialized component type.