Add uninstall functionality
For files and directories this is as simple as removing them. Other
components use the tools used to manage them. Notice that the
status check info for each component is not used. The idea is to
allow the uninstall to fail gracefully if the component is not
installed rather than relying on the status check.