Continued work on order resolution
It is desired, and reasonable, to assume that a method does not
need to be created more than once. Methods may need to be run in a
specific order, but a single method should not need to be run
more than once. Order within a method should also be discovered.