Create seperate functions for each retrieval
The Python script must retrieve the subject line from a specified
cert file and the open id for a specified user. I do not know why
it is this cert file that is being used, only that it was what has
been used. Absolute paths are used to ensure the correct file is
executed, but at the cost of fragility. The installer must install
these components into the correct directory, /esg/config/myproxy