Run startup hooks before cert installation
The ordering of steps here, and likely other places through-out the
install, is EXTREMELY precarious. There is a desperate need to be
more thoughtful when it comes to the order of steps. Also, the
grouping of steps, why are the /etc/grid-security/certifcates being
installed right here at the end? In the startup function? Should it
not be installed with esgf-security? If there were not such a need
to simply 'make it work' this large scale problem could reveive
some TLC. For now, this seems to work.