Add resiliancy to EnvWriter with mirror JSON file
The role of the JSON file is to manage what exports and sources
are already are present in the env file. Appending paths to path
variables such as PATH and LD_LIBRARY_PATH is a common process and
EnvWriter is now capable of both appending and prepending. If the
path variable does not exist, it is created with the specified
value.