Have pip install most components until conda improves install time
conda's install process takes a significant amount of time compared
to pip's install. To improve bootstrap time it is preffered to use
pip to install those packages which are available to it. Also, the
-c cdat specification does not appear to be doing anything and may
be increasing the "Solving Environment" time. Tests of this install
process indicate about a 40-50 second speed up compared to the
previous process.